Create a RecipeCell view (#4)
Creates a RecipeCell, which the main display will be displaying in a
list. It displays all relevant content for recipes, including handling
loading the images with a placeholder.

Note that the cell currently has borders around most of its main views
currently for ease of visual debugging. These will get removed when done
with the rest of the UI.

Also bring in SDWebImageSwiftUI for image loading and caching
simplicity.